refactor: separate core algorithms into modules for imputation, classification, and forecasting models;
