+++ pypots/classification/brits/modules/__init__.py
@@ -1,0 +1,13 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from .core import _BRITS
+
+
+__all__ = [
+    "_BRITS",
+]
