+++ pypots/classification/brits/modules/submodules.py
@@ -19,7 +19,7 @@ from typing import Union
 import torch
 import torch.nn as nn
 
-from ...imputation.brits.model import (
+from pypots.imputation.brits.model import (
     RITS as imputation_RITS,
 )
