+++ pypots/classification/raindrop/modules/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from .core import _Raindrop
+
+__all__ = [
+    "_Raindrop",
+]
